pyproject.toml
[project]
name = "life-wallpaper"
version = "0.1.0"
description = "High-fidelity 4K wallpaper dashboard visualizing life progress"
authors = [{name = "User", email = "user@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
life-wallpaper = "life_wallpaper.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/life_wallpaper"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
